+++ packages/db/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.39.3 - 2025-02-19
+
+### Fixed
+
+- NPostgres: fixed `timeout` opt in `.req`
+
+### Changed
+
+- NPostgres: changed order of parameters in protected `.withTimeout` method.
+
 ## 0.39.2 - 2025-02-19
 
 ### Changed